backend: Added extra to clue and game models
Now we can define the options of a game and a clue in the extra field
and it will be used instead of the setted in the game and clues text.

We should define an "options" field in the json text with all the
options.

This give us the possibility to define the options easily, without the
need to write it in the clues, but with this method we'll show all
options from the first time.